+++ Source/Steam Library Manager/Properties/AssemblyInfo.cs
@@ -51,6 +51,6 @@ using System.Windows;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.5.0.2")]
+[assembly: AssemblyFileVersion("1.5.0.3")]
 [assembly: NeutralResourcesLanguage("en")]